RavenDB-24020 - Going back to using the .NET Thread Pool
All attempts (blocking collection, channels, dedicated threads, etc) has ended up significantly worse in terms of performance.
The design means that we have a max of N batches of work to do (where N is likely small and fixed) and those tend to be very short lived.
So even under load, it isn't expected to fill the thread pool and cause a train.